Migrate Sunburst Renderer to React
