+++ client/app/visualizations/sunburst/Renderer.jsx
@@ -2,13 +2,13 @@ import React, { useState, useEffect, useMemo } from 'react';
 import resizeObserver from '@/services/resizeObserver';
 import { RendererPropTypes } from '@/visualizations';
 
-import initSankey from './initSankey';
+import initSunburst from './initSunburst';
 import './renderer.less';
 
 export default function Renderer({ data }) {
   const [container, setContainer] = useState(null);
 
-  const render = useMemo(() => initSankey(data), [data]);
+  const render = useMemo(() => initSunburst(data), [data]);
 
   useEffect(() => {
     if (container) {
@@ -20,7 +20,7 @@ export default function Renderer({ data }) {
     }
   }, [container, render]);
 
-  return (<div className="sankey-visualization-container" ref={setContainer} />);
+  return (<div className="sunburst-visualization-container" ref={setContainer} />);
 }
 
 Renderer.propTypes = RendererPropTypes;
