+++ client/app/visualizations/sunburst/index.js
@@ -1,42 +1,17 @@
-import { debounce } from 'lodash';
-import Sunburst from '@/lib/visualizations/sunburst';
-import { angular2react } from 'angular2react';
 import { registerVisualization } from '@/visualizations';
 
+import Renderer from './Renderer';
 import Editor from './Editor';
 
-const SunburstSequenceRenderer = {
-  template: '<div class="sunburst-visualization-container" resize-event="handleResize()"></div>',
-  bindings: {
-    data: '<',
-    options: '<',
-  },
-  controller($scope, $element) {
-    const container = $element[0].querySelector('.sunburst-visualization-container');
-    let sunburst = new Sunburst($scope, container);
+export default function init() {
+  registerVisualization({
+    type: 'SUNBURST_SEQUENCE',
+    name: 'Sunburst Sequence',
+    getOptions: options => ({ ...options }),
+    Renderer,
+    Editor,
 
-    function update() {
-      sunburst.remove();
-      sunburst = new Sunburst($scope, container);
-    }
-
-    $scope.handleResize = debounce(update, 50);
-  },
-};
-
-export default function init(ngModule) {
-  ngModule.component('sunburstSequenceRenderer', SunburstSequenceRenderer);
-
-  ngModule.run(($injector) => {
-    registerVisualization({
-      type: 'SUNBURST_SEQUENCE',
-      name: 'Sunburst Sequence',
-      getOptions: options => ({ ...options }),
-      Renderer: angular2react('sunburstSequenceRenderer', SunburstSequenceRenderer, $injector),
-      Editor,
-
-      defaultRows: 7,
-    });
+    defaultRows: 7,
   });
 }
